+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `readMany` to `Service`.
 * Allow `ResponseContext.redirect` to take a `Uri`.
 * Add `Angel.mountController`.
+* Add `Angel.findServiceOf`.
 
 # 2.0.0-alpha.10
 * All calls to `Service.parseId` are now affixed with the `<Id>` argument.